Do not require spec dependencies multiple times
Konacha::SpecsHelper#spec_include_tag is a view helper method for
including specs into views. Under the covers, it uses a SpecBundler to
make sure that no spec dependencies are included twice.

The SpecBundler's job is to provide an interface for adding in specs and
getting all of the assets that those specs rely on. Sprockets doesn't
have an easy built-in way of doing that that does not result in
duplicated assets, so this class is necessary.

Fixes #3.