Display more helpful error messages
When Selenium fails somehow, rake would previously only display "unknown
error" or some such message, whereas now it gives a clue that the
Capybara driver might have failed, and it shows the exception class and
message.

I chose a fairly specific message ("Error communicating with browser
process") because that's the most likely type of error to occur, and I
wanted to clearly mark it as a an environment error, as opposed to an
error in the JavaScript test code.

We have to rescue everything (and not just
Selenium::WebDriver::Error::ServerError) because the user might select a
different Capybara driver.