Run tests in an iframe, with <body id="konacha">
This change is mostly motivated by usability. When you're testing a JS
MVC application, having an iframe makes it much easier to visually
inspect the state of the application, and select specific elements to
look at in the DOM inspector.

I'm hoping that for my team mates that are getting started with the MVC
framework and Konacha, being able to *see* what's going will lessen the
pain of the initial learning curve.

Performance

There seems to be a very slight performance degradation (5-10%) from
painting everything. I feel OK incurring that though, as the iframe
makes Konacha so much more usable.

The delay added by the initial top-level frame is about 20ms on my
system.

Layout

At the moment, if you have a smaller window, it responsively scales the
iframe and the report, but it's all a bit squished for my taste. We have
to leave ample space so error messages don't run behind the iframe. But
perhaps a better way would be to leave a little less space, and add a
"toggle iframe" button to hide it in case an error message runs too
long.

This is perhaps not something that prevents us from merging this, but
I'd like to improve it sooner or later.