Sort the spec_paths to avoid undefined ordering
I think that it's good practice to keep file lists sorted. Otherwise we
can get file-system dependent (undefined) file ordering, and we'll have
order-dependent tests showing up as unreproducible failures on the build
server, which is worse than having them pass silently.

Rails.application.assets seems to sort already, but let's be on the safe
side and treat it like a Dir[] call.