Require test_helper directly in the iframe
With this change, we could massively speed up our entire test suite.
Around a factor of 10 to 20.

We currently introduced one failing spec with this change, it is for discussion,
wether the spec could be changed to adapt our changes in the code (or improvements?)

Since all our tests are still running with this patch, we safely assume, that
everything is still fine.